LF-4598 Removed redirection and made feddback from self-contained
