+++ packages/webapp/src/containers/Help/index.jsx
@@ -1,11 +1,20 @@
 import PureHelpRequestPage from '../../components/Help';
 import { useDispatch, useSelector } from 'react-redux';
 import { supportFileUpload } from './saga';
-import { isHelpLoadingSelector, startSendHelp } from '../Home/homeSlice';
+import {
+  isHelpLoadingSelector,
+  startSendHelp,
+  dismissHelpRequestModal,
+  showHelpRequestModalSelector,
+} from '../Home/homeSlice';
 import { userFarmSelector } from '../userFarmSlice';
+import RequestConfirmationComponent from '../../components/Modals/RequestConfirmationModal';
 
 export default function HelpRequest({ closeDrawer }) {
   const dispatch = useDispatch();
+
+  const showHelpRequestModal = useSelector(showHelpRequestModalSelector);
+  const showRequestConfirmationModalOnClick = () => dispatch(dismissHelpRequestModal());
 
   const handleSubmit = (file, data, resetForm) => {
     dispatch(startSendHelp());
@@ -27,12 +36,20 @@ export default function HelpRequest({ closeDrawer }) {
   const { email, phone_number } = useSelector(userFarmSelector);
   const loading = useSelector(isHelpLoadingSelector);
   return (
-    <PureHelpRequestPage
-      onSubmit={handleSubmit}
-      onCancel={onCancel}
-      email={email}
-      phoneNumber={phone_number}
-      isLoading={loading}
-    />
+    <>
+      <PureHelpRequestPage
+        onSubmit={handleSubmit}
+        onCancel={onCancel}
+        email={email}
+        phoneNumber={phone_number}
+        isLoading={loading}
+      />
+      {showHelpRequestModal && (
+        <RequestConfirmationComponent
+          onClick={showRequestConfirmationModalOnClick}
+          dismissModal={showRequestConfirmationModalOnClick}
+        />
+      )}
+    </>
   );
 }
